# Improve head chunk allocations when reading samples.